deps: bump promql-parser to 0.9 for limitk grammar
v0.9.0 adds grammar rules for limitk and limit_ratio in the
aggregate_op production. v0.8.0 declared these as `%expect-unused`
tokens but never bound them in the grammar, so `limitk(...)` queries
failed at parse time before this bump.

The bump also widens the token id type from u8 to u16. Updated two
local helper signatures (aggregate_op_name, binary_op_kind) to match.
No behavior change.

+++ timeseries/src/promql/plan/lowering.rs
@@ -770,7 +770,7 @@ fn aggregate_streaming(
     }
 }
 
-fn aggregate_op_name(op_id: u8) -> &'static str {
+fn aggregate_op_name(op_id: u16) -> &'static str {
     match op_id {
         T_SUM => "sum",
         T_AVG => "avg",
@@ -832,7 +832,7 @@ fn lower_binary(bin: &parser::BinaryExpr, ctx: &LoweringContext) -> Result<Logic
     })
 }
 
-fn binary_op_kind(op_id: u8, return_bool: bool) -> Option<BinaryOpKind> {
+fn binary_op_kind(op_id: u16, return_bool: bool) -> Option<BinaryOpKind> {
     Some(match op_id {
         T_ADD => BinaryOpKind::Add,
         T_SUB => BinaryOpKind::Sub,
